rename subnet_tracker to subnet_service and move get_committee_info_for_subnet there
