+++ Cargo.toml
@@ -25,7 +25,7 @@ members = [
     "anchor/processor",
     "anchor/qbft_manager",
     "anchor/signature_collector",
-    "anchor/subnet_tracker",
+    "anchor/subnet_service",
     "anchor/validator_store",
 ]
 
@@ -59,7 +59,7 @@ qbft_manager = { path = "anchor/qbft_manager" }
 signature_collector = { path = "anchor/signature_collector" }
 ssv_network_config = { path = "anchor/common/ssv_network_config" }
 ssv_types = { path = "anchor/common/ssv_types" }
-subnet_tracker = { path = "anchor/subnet_tracker" }
+subnet_service = { path = "anchor/subnet_service" }
 version = { path = "anchor/common/version" }
 
 beacon_node_fallback = { git = "https://github.com/sigp/lighthouse", rev = "9b84dac" }
