+++ anchor/message_sender/Cargo.toml
@@ -14,6 +14,6 @@ openssl = { workspace = true }
 processor = { workspace = true }
 slot_clock = { workspace = true }
 ssv_types = { workspace = true }
-subnet_tracker = { workspace = true }
+subnet_service = { workspace = true }
 tokio = { workspace = true }
 tracing = { workspace = true }
