+++ anchor/message_sender/src/impostor.rs
@@ -1,5 +1,5 @@
 use ssv_types::{CommitteeId, consensus::UnsignedSSVMessage, message::SignedSSVMessage};
-use subnet_tracker::SubnetId;
+use subnet_service::SubnetId;
 use tokio::sync::mpsc;
 use tracing::debug;
