+++ anchor/message_sender/src/network.rs
@@ -13,7 +13,7 @@ use ssv_types::{
     CommitteeId, OperatorId, consensus::UnsignedSSVMessage, message::SignedSSVMessage,
 };
 use ssz::Encode;
-use subnet_tracker::SubnetId;
+use subnet_service::SubnetId;
 use tokio::sync::{mpsc, mpsc::error::TrySendError};
 use tracing::{debug, error, warn};
