+++ anchor/network/Cargo.toml
@@ -6,7 +6,6 @@ authors = ["Sigma Prime <contact@sigmaprime.io>"]
 
 [dependencies]
 async-trait = "0.1.85"
-database = { workspace = true }
 dirs = { workspace = true }
 discv5 = { workspace = true }
 ethereum_ssz = { workspace = true }
@@ -34,7 +33,7 @@ serde = { workspace = true }
 serde_json = "1.0.137"
 ssv_types = { workspace = true }
 ssz_types = "0.10"
-subnet_tracker = { workspace = true }
+subnet_service = { workspace = true }
 task_executor = { workspace = true }
 thiserror = { workspace = true }
 tokio = { workspace = true }
