+++ anchor/subnet_service/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "subnet_tracker"
+name = "subnet_service"
 version = "0.1.0"
 edition = { workspace = true }
 authors = ["Sigma Prime <contact@sigmaprime.io>"]
